% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoStopCluster.R
\name{autoStopCluster}
\alias{autoStopCluster}
\title{Sub-function to auto-stop the clusters created for parallel processes}
\usage{
autoStopCluster(cl)
}
\arguments{
\item{cl}{The cluster created by the makePSOCKcluster function}
}
\description{
Sub-function to auto-stop the clusters created for parallel processes
}
\keyword{internal}
