% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_genes_qtls_around_markers.R
\name{find_genes_qtls_around_markers}
\alias{find_genes_qtls_around_markers}
\title{Search genes and QTLs around candidate regions}
\usage{
find_genes_qtls_around_markers(db_file, marker_file, method = c("gene",
  "qtl"), marker = c("snp", "haplotype"), interval = 0,
  nThreads = NULL)
}
\arguments{
\item{db_file}{file with the gene mapping or QTL information. For the gene mapping, you should use the .gtf file download from Ensembl data base. For the QTL search, you need to inform the .gff file that can be downloaded from Animal QTlLdb.}

\item{marker_file}{The file with the SNP or haplotype positions. Detail: For SNP files, you must have a column called “CHR” and a column called “BP” with the chromosome and base pair position, respectively. For the haplotype, you must have three columns: “CHR”, “BP1” and “BP2”. All the columns names are capitals.}

\item{method}{“gene” or “qtl”}

\item{marker}{"snp" or "haplotype"}

\item{interval}{The interval in base pair}

\item{nThreads}{Number of threads to be used}
}
\value{
A dataframe with the genes or QTLs mapped within the specified intervals
}
\description{
Takes a list of candidate markers and or regions (haplotypes) and search for genes or QTLs in a determined interval
}
