% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_markers_genes.R
\name{find_markers_genes}
\alias{find_markers_genes}
\title{Function to annotate markers and the respective p-values within genes}
\usage{
find_markers_genes(db_file, marker_file, int = 0)
}
\arguments{
\item{db_file}{A data frame obtained from the import_gff_gtf containing the gtf information}

\item{marker_file}{A data frame with the results of the association test performed for each marker}

\item{int}{The interval (in base pairs) used to annotated markers downstream and upstream from the genes coordinates}
}
\value{
A data frame containing the markers mapped within the selected interval for each gene in the annotation file
}
\description{
Function to annotate markers and the respective p-values within genes
}
\keyword{internal}
