% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inferCut.R
\name{inferCut}
\alias{inferCut}
\title{Compute the eigenvalues for the linkage disequilibrium (LD) matrix and use as input for PCA_Meff function to compute the effective number of markers}
\usage{
inferCut(mat.r, cut.off)
}
\arguments{
\item{mat.r}{A matrix composed by the LD between markers}

\item{cut.off}{The threshold for percentage of the sum of the variances explained by the markers}
}
\value{
The effective number of markers identified by the SimpleM approach
}
\description{
Compute the eigenvalues for the linkage disequilibrium (LD) matrix and use as input for PCA_Meff function to compute the effective number of markers
}
\keyword{internal}
