% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_LD.R
\name{meta_LD}
\alias{meta_LD}
\title{unction to perform Meta-Analysis with LD Correlation Coefficients}
\usage{
meta_LD(marker_ld, marker_pvalues)
}
\arguments{
\item{marker_ld}{A dataframe containing the pairwise linkage disequilibrium between markers in a chromosome}

\item{marker_pvalues}{A vector with the p-values for the SNPs annotated within each gene}
}
\value{
A vector of p-values for each gene annotated within the defined coordinates
}
\description{
unction to perform Meta-Analysis with LD Correlation Coefficients
}
\keyword{internal}
