% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiCores.R
\name{MultiCores}
\alias{MultiCores}
\title{Sub-function to allow multithreading}
\usage{
MultiCores(nThreads)
}
\arguments{
\item{nThreads}{The number of threads to be used}

\item{int}{The interval in base pair}
}
\value{
Registers the parallel work in the required cores
}
\description{
Takes the number of cores to be used and check if it is possible to allocate the work in the in the respective cores
}
\keyword{internal}
