% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Nseg_chr.R
\name{Nseg_chr}
\alias{Nseg_chr}
\title{Estimate the number of independent segments in a chromosome based on the effective population size}
\usage{
Nseg_chr(chr.table, chr_length, Ne)
}
\arguments{
\item{chr.table}{A table containing the chromosomes and the chromosomal length (in centiMorgans).}

\item{chr_length}{The name of the column where the length of the chromosomes are informed.}

\item{Ne}{The effective population size.}
}
\value{
A data frame with the effevtive number of segments in each chromosome.
}
\description{
Estimate the number of independent segments in a chromosome based on the effective population size
}
\details{
This function uses a adapted version of the formula proposed by Goddard et al. (2011) to estimate the independent number of segments in a chromosome based on the effective population size.
}
\references{
Goddard et al. (2011) Journal of animal breeding and genetics, Volume 128, Issue 6, Pages 409-421.
(\doi{10.1111/j.1439-0388.2011.00964.x})
}
