% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCA_Meff.R
\name{PCA_Meff}
\alias{PCA_Meff}
\title{Compute Meff statistic based on PCA to determine the number of effective markers}
\usage{
PCA_Meff(eigenV, cut.off)
}
\arguments{
\item{eigenV}{The eigenvalues obtained from the linkage disequilibrium matrix}

\item{cut.off}{The threshold for percentage of the sum of the variances explained by the markers}
}
\value{
The effective number of markers identified by the SimpleM approach
}
\description{
Compute Meff statistic based on PCA to determine the number of effective markers
}
\keyword{internal}
