% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PleioChiTest.R
\name{PleioChiTest}
\alias{PleioChiTest}
\title{Compute a multi-trait test statistic for pleiotropic effects using summary statistics from association tests}
\usage{
PleioChiTest(data)
}
\arguments{
\item{data}{A data frame with the first column containing the SNP name and the remaining columns the signed t-values obtained for each marker in the association studies individually performed for each trait.}
}
\value{
A data frame with the multi-trait chi-squared statistics and the correspondent p-value obtained for each SNP.
}
\description{
Compute a multi-trait test statistic for pleiotropic effects using summary statistics from association tests
}
\details{
This function tests a null hypothesis stating that each SNP does not affect any of the traits included in the input file. The method applied here is an implementation of the statistic proposed at Bolormaa et al. (2014) and is approximately distributed as a chi-squared with n degrees of freedom, where n is equal the number of traits included in the input file.
}
\references{
Bolormaa et al. (2014) Plos Genetics, Volume 10, Issue 3, e1004198.
(\doi{10.1371/journal.pgen.1004198})
}
