% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QTLenrich_plot.R
\name{QTLenrich_plot}
\alias{QTLenrich_plot}
\title{Plot enrichment results for QTL enrichment analysis}
\usage{
QTLenrich_plot(qtl_enrich, x, pval)
}
\arguments{
\item{qtl_enrich}{The output from qtl_enrich function}

\item{x}{Id column to be used from the qtl_enrich output}

\item{pval}{P-value to be used in the plot. If "p_value" informed, a non-adjusted pvalue will be plot. If "padj" informed, the adjusted p-value from the qtl enrichment analysis will be plotted.}
}
\description{
Takes the output from overlapping_amoung_groups function and creates a heatmap with the overlapping between groups
}
\examples{
data(QTLwindows)
\donttest{qtl.out <- find_genes_qtls_around_markers(db_file="QTL_db.gff",
marker_file=QTLwindows,method="qtl",
marker="haplotypes",interval=100000)}
\donttest{qtl.enrich.out<-qtl_enrich(qtl_db="QTL_db.gff",
qtl_file=qtl.out,qtl_type="QTL_type",
enrich_type="genome",chr.subset=NULL,
n.it=1000,padj="fdr")}
\donttest{QTLenrich_plot(qtl_enrich=qtl.enrich.out,x="QTL_type",pval="adjpval")}
}
