% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relationship_plot.R
\name{relationship_plot}
\alias{relationship_plot}
\title{Plot relationship between data and grouping factors}
\usage{
relationship_plot(
  qtl_file,
  x,
  y,
  grid.col = "gray60",
  degree = 90,
  canvas.xlim = c(-2, 2),
  canvas.ylim = c(-2, 2),
  cex,
  gap
)
}
\arguments{
\item{qtl_file}{The output from find_genes_qtls_around_markers function}

\item{x}{The first grouping factor, to be plotted in the left hand side of the chord plot}

\item{y}{The second grouping factor, to be plotted in the left hand side of the chord plot}

\item{grid.col}{A character with the grid color for the chord plot or a vector with different colors to be used in the grid colors. Note that when a color vector is provided, the lenght of this vector must be equal the number of sectors in the chord plot}

\item{degree}{A numeric value corresponding to the starting degree from which the circle begins to draw. Note this degree is always reverse-clockwise}

\item{canvas.xlim}{The coordinate for the canvas in the x-axis. By default is c(-1,1)}

\item{canvas.ylim}{The coordinate for the canvas in the y-axis. By default is c(-1,1)}

\item{cex}{The size of the labels to be printed in the plot}

\item{gap}{A numeric value corresponding to the gap between the chord sectors}
}
\value{
A chords relating x and y
}
\description{
Takes the output from find_genes_qtls_around_markers function and creates a chord plot with the relationship between groups
}
\examples{
data(QTLwindows)
\donttest{genes.out <- find_genes_qtls_around_markers(db_file="gene.gtf",marker_file=QTLwindows,method="gene",
marker="haplotypes",interval=100000)}
\donttest{relationship_plot(qtl_file=genes.out,x="Reference",
y="gene_id")}
}
