% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resampFun_trait.R
\name{resampFun.trait}
\alias{resampFun.trait}
\title{Function to performn resampling from qtl output}
\usage{
resampFun.trait(sub.qtl, i)
}
\arguments{
\item{sub.qtl}{temporary qtl file splitted by chromosome created by enrich_qtl function}

\item{i}{number of interactions to be used during the bootstrap analysis}
}
\description{
Takes a list of candidate markers and search for genes a determined interval
}
