% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitQTL_comment.R
\name{splitQTL_comment}
\alias{splitQTL_comment}
\title{Sub-function to split comment column from QTL output}
\usage{
splitQTL_comment(output_qtls)
}
\arguments{
\item{output_qtls}{Output from QTL annotation}
}
\description{
Takes a list of candidate markers and search for genes a determined interval
}
