% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sub_qtlEnrich_geno.R
\name{sub_qtlEnrich_geno}
\alias{sub_qtlEnrich_geno}
\title{Performs a QTL enrichment analysis based in a Bootstrap simulation for each QTL class using the QTL information across the whole genome}
\usage{
sub_qtlEnrich_geno(
  qtl_file,
  qtl.file.types,
  table.qtl,
  qtl_db,
  parallel,
  n.it,
  nThreads
)
}
\arguments{
\item{qtl_file}{The output from find_genes_qtls_around_markers function}

\item{qtl.file.types}{A vector with the observed QTL classes}
}
\value{
A data frame with the p-value for th enrichment result
}
\description{
Takes the output from find_genes_qtls_around_markers and run a QTL enrichment analysis
}
\keyword{internal}
