% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WZ_ld.R
\name{WZ_ld}
\alias{WZ_ld}
\title{Function to perform Weighted Z-score Approach with LD Information}
\usage{
WZ_ld(marker_ld, marker_pvalues)
}
\arguments{
\item{marker_ld}{A data frame containing the pairwise linkage disequilibrium between markers in a chromosome}

\item{marker_pvalues}{A vector with the p-values for the SNPs annotated within each gene}
}
\value{
A vector of p-values for each gene annotated within the defined coordinates
}
\description{
Function to perform Weighted Z-score Approach with LD Information
}
\keyword{internal}
