% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qtl_enrich.R
\name{qtl_enrich}
\alias{qtl_enrich}
\title{Performs a QTL enrichment analysis based in a Bootstrap simulation for each QTL class}
\usage{
qtl_enrich(qtl_db, qtl_file, qtl_type = c("QTL_type", "trait"),
  enrich_type = c("genome", "chromosome"), chr.subset = NULL,
  n.it = NULL, nThreads = NULL, padj = c("holm", "hochberg",
  "hommel", "bonferroni", "BH", "BY", "fdr", "none"), parallel = c("no",
  "multicore", "snow"))
}
\arguments{
\item{qtl_db}{The .gff file that can be downloaded from Animal QTlLdb}

\item{qtl_file}{The output from find_genes_qtls_around_markers function}

\item{qtl_type}{A character indicating which type of enrichment will be performed. QTL_type indicates that the enrichment processes will be performed for the QTL classes, while trait indicates that the enrichment analysis will be performed for each trait individually.}

\item{enrich_type}{A character indicating if the enrichment analysis will be performed for all the chromosomes ("genome") or for a subset of chromosomes ("chromosome). If the "genome" option is selected, the results reported are the merge of all chromosomes.}

\item{chr.subset}{If enrich_type equal "chromosome", it is possible to define a subset of chromosomes to be analyzed. The default is equal NULL. Therefore, all the chromosomes will be analyzed.}

\item{n.it}{number of iterations for the bootstrap simulation.}

\item{nThreads}{The number of threads used.}

\item{padj}{The alogorithm for multiple testing correction to be adopted ("holm", "hochberg", "hommel", "bonferroni", "BH", "BY","fdr", "none").}

\item{parallel}{The type of parallel operation to be used.}
}
\value{
A data frame with the p-value for th enrichment result
}
\description{
Takes the output from find_genes_qtls_around_markers and run a QTL enrichment analysis
}
