% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QTLenrich_plot.R
\name{QTLenrich_plot}
\alias{QTLenrich_plot}
\title{Plot enrichment results for QTL enrichment analysis}
\usage{
QTLenrich_plot(qtl_enrich, x, pval)
}
\arguments{
\item{qtl_enrich}{The output from qtl_enrich function}

\item{x}{Id column to be used from the qtl_enrich output}

\item{pval}{P-value to be used in the plot. The name informed to this argument must match the p-value column name in the enrichment table}
}
\value{
A plot with the QTL enrichment results
}
\description{
Takes the output from qtl_enrich function and creates a bubble plot with enrichment results
}
