% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QTLenrich_plot.R
\name{QTLenrich_plot}
\alias{QTLenrich_plot}
\title{Plot enrichment results for QTL enrichment analysis}
\usage{
QTLenrich_plot(qtl_enrich, x, pval)
}
\arguments{
\item{qtl_enrich}{The output from qtl_enrich function}

\item{x}{Id column to be used from the qtl_enrich output}

\item{pval}{P-value to be used in the plot. If "p_value" informed, a non-adjusted pvalue will be plot. If "padj" informed, the adjusted p-value from the qtl enrichment analysis will be plotted.}
}
\description{
Takes the output from overlapping_amoung_groups function and creates a heatmap with the overlapping between groups
}
