% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitQTL_comment.R
\name{splitQTL_comment}
\alias{splitQTL_comment}
\title{Sub-function to split comment column from QTL output}
\usage{
splitQTL_comment(output.final)
}
\arguments{
\item{output.final}{Output from QTL annotation}
}
\value{
A data frame with the extra_info column content, from the gff file, broken in several additional columns
}
\description{
Takes a list of candidate markers and search for genes a determined interval
}
\keyword{internal}
