% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sub_qtlEnrich_chrom.R
\name{sub_qtlEnrich_chrom}
\alias{sub_qtlEnrich_chrom}
\alias{sub_qtlEnrich_chom}
\title{Performs a QTL enrichment analysis based in a Bootstrap simulation for each QTL class using the QTL information per chromosome}
\usage{
sub_qtlEnrich_chom(
  qtl_file,
  qtl_type,
  qtl.file.types,
  table.qtl.class,
  padj,
  qtl_db,
  search_qtl,
  nThreads
)
}
\arguments{
\item{qtl_file}{The output from find_genes_qtls_around_markers function}

\item{qtl_type}{A string indicating with QTL enrichment will be performed: "QTL_type" or "Name"}

\item{qtl.file.types}{A vector with the observed QTL classes}

\item{table.qtl.class}{An frequency table for the number of each QTL in each chromosome}

\item{qtl_db}{The QTL annotation database}

\item{search_qtl}{The column to perform the QTL searching in counting from the QTL annotation database}

\item{nThreads}{Number of threads for parallel processing}
}
\value{
A data frame with the p-value for th enrichment result
}
\description{
Takes the output from find_genes_qtls_around_markers and run a QTL enrichment analysis
}
\keyword{internal}
